Separate user API and some test APIs from back-end
