+++ bot/src/main/kotlin/org/epilink/bot/LinkServerEnvironment.kt
@@ -21,10 +21,10 @@ import org.epilink.bot.discord.*
 import org.epilink.bot.discord.LinkDiscordMessagesImpl
 import org.epilink.bot.discord.LinkRoleManagerImpl
 import org.epilink.bot.http.*
-import org.epilink.bot.http.endpoints.LinkMetaApi
+import org.epilink.bot.http.endpoints.*
 import org.epilink.bot.http.endpoints.LinkMetaApiImpl
-import org.epilink.bot.http.endpoints.LinkRegistrationApi
 import org.epilink.bot.http.endpoints.LinkRegistrationApiImpl
+import org.epilink.bot.http.endpoints.LinkUserApiImpl
 import org.koin.core.context.startKoin
 import org.koin.core.logger.Level
 import org.koin.dsl.module
@@ -107,6 +107,8 @@ class LinkServerEnvironment(
         single<LinkRegistrationApi> { LinkRegistrationApiImpl() }
 
         single<LinkMetaApi> { LinkMetaApiImpl() }
+
+        single<LinkUserApi> { LinkUserApiImpl() }
     }
 
     /**
